requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
telethon==1.34.0
python-telegram-bot==20.7
openai==1.3.7
redis==5.0.1
pydantic==2.5.0
python-dotenv==1.0.0
apscheduler==3.10.4
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
loguru==0.7.2
pytz==2023.3
psycopg2-binary==2.9.10
